% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/azmpdata-package.R
\docType{package}
\name{azmpdata}
\alias{azmpdata}
\alias{azmp}
\title{azmpdata - manage local data of DFO-MPO calanus database}
\description{
Functions for managing local repository of DFO-MPO calanus database.
}
