% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cols_calanus}
\alias{cols_calanus}
\title{Generate a listing of column types suitable for\code{\link[readr]{read_tsv}}}
\usage{
cols_calanus()
}
\value{
a named list of column types
}
\description{
Generate a listing of column types suitable for\code{\link[readr]{read_tsv}}
}
