% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{cols_ctd_prof}
\alias{cols_ctd_prof}
\title{Retrieve the columns to select when reading a prof ctd}
\usage{
cols_ctd_prof()
}
\value{
charcater vector
}
\description{
Retrieve the columns to select when reading a prof ctd
}
