% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ctd}
\alias{read_ctd}
\title{Read a single CTD - note that this is sensitive to prof and surf}
\usage{
read_ctd(
  filename = get_data_path("Maritimes", "a_MEDS_profile_prof_2015_2020.csv"),
  form = c("tibble", "sf")[1]
)
}
\arguments{
\item{filename}{character, the file to read}

\item{form}{character, one of 'tibble' or 'sf'}
}
\value{
tibble or sf POINT table
}
\description{
Read a single CTD - note that this is sensitive to prof and surf
}
