% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ctd_prof}
\alias{read_ctd_prof}
\title{Read a single \code{prof} ctd file}
\usage{
read_ctd_prof(
  filename,
  data_type = c("CD", "CU"),
  source_id = c("BIO", "NAFC"),
  q_date_time = c(1, 2, 5),
  q_pos = c(1, 2, 5),
  q_temp = c(1, 2, 5),
  dp_flag = c(1, 2, 5),
  q_psal = c(1, 2, 5)
)
}
\arguments{
\item{filename}{char, the file to read}

\item{data_type}{char, codes of data types to retain}

\item{source_id}{char, codes of data sources to retain}

\item{q_date_time}{num, codes of q_data_time to retain}

\item{q_pos}{num, codes of q_pos to retain}

\item{q_temp}{num, codes of q_temp to retain}

\item{dp_flag}{num, codes of dp_flag to retain}

\item{q_psal}{num, codes of q_psal to retain}
}
\value{
tibble
}
\description{
Read a single \code{prof} ctd file
}
